request #12648: Tuleap Enterprise plugins should be buildable for CentOS/RHEL 7
Packages can be built for RHEL/CentOS 7 this way:
    TULEAP_PATH=~/Workspace/tuleap/ WORKSPACE=/tmp OS='centos:7' ./build-rpm.sh

+++ plugins/timetracking/build-rpm.sh
@@ -12,7 +12,12 @@ if [ ! -d "$WORKSPACE" ]; then
     exit 1
 fi
 
-DOCKERIMAGE=build-plugin-timetracking
+if [[ -z "$OS" ]]; then
+    echo "*** ERROR: OS is missing"
+    exit 1
+fi
 
-docker build -t "$DOCKERIMAGE" rpm
+DOCKERIMAGE=build-plugin-timetracking-"$OS"
+
+docker build --build-arg OS="$OS" -t "$DOCKERIMAGE" rpm
 docker run --rm -e "RELEASE=$RELEASE" -v "$TULEAP_PATH":/tuleap:ro -v "$WORKSPACE":/output -e UID="$(id -u)" -e GID="$(id -g)" "$DOCKERIMAGE"
